# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KAYGO API - Plateforme de mise en relation logistique France-Martinique
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  traveler: "traveler",
  sender: "sender",
  receiver: "receiver",
  admin: "admin",
} as const;

export interface RegisterRequest {
  firstName: string;
  lastName: string;
  email: string;
  phone?: string;
  password: string;
  role: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  role: string;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  avatarUrl?: string | null;
  verificationStatus: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserResponse {
  user: User;
}

export interface UsersListResponse {
  users: User[];
  total: number;
}

export interface UpdateUserRequest {
  firstName?: string;
  lastName?: string;
  phone?: string;
  avatarUrl?: string;
  verificationStatus?: string;
}

export type VerifyRequestStatus =
  (typeof VerifyRequestStatus)[keyof typeof VerifyRequestStatus];

export const VerifyRequestStatus = {
  verified: "verified",
  rejected: "rejected",
} as const;

export interface VerifyRequest {
  status: VerifyRequestStatus;
  note?: string;
}

export interface Trip {
  id: number;
  travelerId: number;
  departureCity: string;
  arrivalCity: string;
  departureDate: string;
  arrivalDate: string;
  flightNumber?: string | null;
  transportType?: string;
  baggageTotalKg: number;
  baggageUsedKg: number;
  baggageFreeKg: number;
  baggageType?: string;
  acceptsExtraBaggage?: boolean;
  acceptedCategories?: string[];
  minReward: number;
  notes?: string | null;
  status: string;
  createdAt: string;
  travelerName?: string | null;
}

export interface TripResponse {
  trip: Trip;
}

export interface TripsListResponse {
  trips: Trip[];
  total: number;
}

export interface CreateTripRequest {
  travelerId: number;
  departureCity: string;
  arrivalCity: string;
  departureDate: string;
  arrivalDate: string;
  flightNumber?: string;
  transportType?: string;
  baggageTotalKg: number;
  baggageUsedKg: number;
  baggageFreeKg: number;
  baggageType?: string;
  acceptsExtraBaggage?: boolean;
  acceptedCategories?: string[];
  minReward: number;
  notes?: string;
}

export interface UpdateTripRequest {
  status?: string;
  baggageUsedKg?: number;
  baggageFreeKg?: number;
  notes?: string;
}

export interface Shipment {
  id: number;
  senderId: number;
  receiverName: string;
  receiverPhone: string;
  departureCity: string;
  arrivalCity: string;
  weightKg: number;
  dimensions?: string | null;
  category: string;
  description?: string | null;
  declaredValue?: number | null;
  photoUrls?: string[];
  urgencyLevel?: string;
  pickupOption?: boolean;
  deliveryOption?: boolean;
  pickupAddress?: string | null;
  deliveryAddress?: string | null;
  status: string;
  createdAt: string;
  senderName?: string | null;
}

export interface ShipmentResponse {
  shipment: Shipment;
}

export interface ShipmentsListResponse {
  shipments: Shipment[];
  total: number;
}

export interface CreateShipmentRequest {
  senderId: number;
  receiverName: string;
  receiverPhone: string;
  departureCity: string;
  arrivalCity: string;
  weightKg: number;
  dimensions?: string;
  category: string;
  description?: string;
  declaredValue?: number;
  photoUrls?: string[];
  urgencyLevel?: string;
  pickupOption?: boolean;
  deliveryOption?: boolean;
  pickupAddress?: string;
  deliveryAddress?: string;
}

export interface UpdateShipmentRequest {
  status?: string;
  photoUrls?: string[];
  notes?: string;
}

export interface Match {
  id: number;
  shipmentId: number;
  tripId: number;
  proposedReward: number;
  platformFee: number;
  serviceFee: number;
  pickupFee: number;
  deliveryFee: number;
  totalPrice: number;
  adminStatus: string;
  travelerStatus: string;
  senderStatus: string;
  createdAt: string;
  shipment?: Shipment | null;
  trip?: Trip | null;
}

export interface MatchResponse {
  match: Match;
}

export interface MatchesListResponse {
  matches: Match[];
  total: number;
}

export interface CreateMatchRequest {
  shipmentId: number;
  tripId: number;
  proposedReward: number;
  platformFee: number;
  serviceFee: number;
  pickupFee: number;
  deliveryFee: number;
  totalPrice: number;
}

export interface UpdateMatchRequest {
  adminStatus?: string;
  travelerStatus?: string;
  senderStatus?: string;
}

export interface Payment {
  id: number;
  shipmentId: number;
  payerId: number;
  amount: number;
  paymentStatus: string;
  payoutStatus: string;
  stripePaymentIntentId?: string | null;
  createdAt: string;
}

export interface PaymentResponse {
  payment: Payment;
}

export interface PaymentsListResponse {
  payments: Payment[];
  total: number;
}

export interface Proof {
  id: number;
  shipmentId: number;
  proofType: string;
  photoUrl?: string | null;
  note?: string | null;
  deliveryCode?: string | null;
  createdAt: string;
}

export interface ProofResponse {
  proof: Proof;
}

export interface ProofsListResponse {
  proofs: Proof[];
}

export interface CreateProofRequest {
  shipmentId: number;
  proofType: string;
  photoUrl?: string;
  note?: string;
  deliveryCode?: string;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  title: string;
  message: string;
  status: string;
  createdAt: string;
}

export interface NotificationsListResponse {
  notifications: Notification[];
}

export type PriceEstimateRequestServiceLevel =
  (typeof PriceEstimateRequestServiceLevel)[keyof typeof PriceEstimateRequestServiceLevel];

export const PriceEstimateRequestServiceLevel = {
  eco: "eco",
  confort: "confort",
  premium: "premium",
} as const;

export interface PriceEstimateRequest {
  weightKg: number;
  pickupOption?: boolean;
  deliveryOption?: boolean;
  urgencyLevel?: string;
  serviceLevel: PriceEstimateRequestServiceLevel;
}

export type PriceEstimateResponseBreakdown = {
  label?: string;
  amount?: number;
  [key: string]: unknown;
};

export interface PriceEstimateResponse {
  transportFee: number;
  serviceFee: number;
  pickupFee: number;
  deliveryFee: number;
  totalPrice: number;
  serviceLevel: string;
  estimatedDays: number;
  breakdown?: PriceEstimateResponseBreakdown;
}

export type AdminStatsResponseRecentActivityItem = {
  type: string;
  message: string;
  time: string;
};

export interface AdminStatsResponse {
  totalTrips: number;
  totalShipments: number;
  validatedMatches: number;
  platformRevenue: number;
  travelerGains: number;
  collectionsToday: number;
  deliveriesToday: number;
  openDisputes: number;
  pendingValidations: number;
  recentActivity: AdminStatsResponseRecentActivityItem[];
}

export interface CreateAdminActionRequest {
  adminId: number;
  targetType: string;
  targetId: number;
  action: string;
  note?: string;
}

export type AdminActionResponseAction = {
  id: number;
  adminId: number;
  targetType: string;
  targetId: number;
  action: string;
  note?: string | null;
  createdAt: string;
};

export interface AdminActionResponse {
  action: AdminActionResponseAction;
}

export type ListUsersParams = {
  role?: string;
  verificationStatus?: string;
};

export type ListTripsParams = {
  status?: string;
  departureCity?: string;
  arrivalCity?: string;
};

export type ListShipmentsParams = {
  status?: string;
  senderId?: number;
};

export type ListMatchesParams = {
  adminStatus?: string;
};